# Support Lua scripts without SHA-1 usage for FIPS compliancy