--- a/lua_test.go
+++ b/lua_test.go
@@ -75,6 +75,37 @@ func TestNewLuaScript(t *testing.T) {
 	}
 }
 
+func TestNewLuaScriptNoSha(t *testing.T) {
+	defer ShouldNotLeaked(SetupLeakDetection())
+	body := strconv.Itoa(rand.Int())
+	sum := sha1.Sum([]byte(body))
+	sha := hex.EncodeToString(sum[:])
+
+	k := []string{"1", "2"}
+	a := []string{"3", "4"}
+
+	c := &client{
+		BFn: func() Builder {
+			return cmds.NewBuilder(cmds.NoSlot)
+		},
+		DoFn: func(ctx context.Context, cmd Completed) (resp RedisResult) {
+			if reflect.DeepEqual(cmd.Commands(), []string{"EVALSHA", sha, "2", "1", "2", "3", "4"}) {
+				t.Fatal("EVALSHA must not be called")
+			}
+			if reflect.DeepEqual(cmd.Commands(), []string{"EVAL", body, "2", "1", "2", "3", "4"}) {
+				return newResult(RedisMessage{typ: '_'}, nil)
+			}
+			return newResult(strmsg('+', "unexpected"), nil)
+		},
+	}
+
+	script := NewLuaScriptNoSha(body)
+
+	if err, ok := IsRedisErr(script.Exec(context.Background(), c, k, a).Error()); ok && !err.IsNil() {
+		t.Fatalf("ret mistmatch")
+	}
+}
+
 func TestNewLuaScriptReadOnly(t *testing.T) {
 	defer ShouldNotLeaked(SetupLeakDetection())
 	body := strconv.Itoa(rand.Int())
@@ -103,6 +134,37 @@ func TestNewLuaScriptReadOnly(t *testing.T) {
 	}
 
 	script := NewLuaScriptReadOnly(body)
+
+	if err, ok := IsRedisErr(script.Exec(context.Background(), c, k, a).Error()); ok && !err.IsNil() {
+		t.Fatalf("ret mistmatch")
+	}
+}
+
+func TestNewLuaScriptReadOnlyNoSha(t *testing.T) {
+	defer ShouldNotLeaked(SetupLeakDetection())
+	body := strconv.Itoa(rand.Int())
+	sum := sha1.Sum([]byte(body))
+	sha := hex.EncodeToString(sum[:])
+
+	k := []string{"1", "2"}
+	a := []string{"3", "4"}
+
+	c := &client{
+		BFn: func() Builder {
+			return cmds.NewBuilder(cmds.NoSlot)
+		},
+		DoFn: func(ctx context.Context, cmd Completed) (resp RedisResult) {
+			if reflect.DeepEqual(cmd.Commands(), []string{"EVALSHA_RO", sha, "2", "1", "2", "3", "4"}) {
+				t.Fatal("EVALSHA_RO must not be called")
+			}
+			if reflect.DeepEqual(cmd.Commands(), []string{"EVAL_RO", body, "2", "1", "2", "3", "4"}) {
+				return newResult(RedisMessage{typ: '_'}, nil)
+			}
+			return newResult(strmsg('+', "unexpected"), nil)
+		},
+	}
+
+	script := NewLuaScriptReadOnlyNoSha(body)
 
 	if err, ok := IsRedisErr(script.Exec(context.Background(), c, k, a).Error()); ok && !err.IsNil() {
 		t.Fatalf("ret mistmatch")
@@ -163,6 +225,36 @@ func TestNewLuaScriptExecMulti(t *testing.T) {
 	}
 }
 
+func TestNewLuaScriptExecMultiNoSha(t *testing.T) {
+	defer ShouldNotLeaked(SetupLeakDetection())
+	body := strconv.Itoa(rand.Int())
+
+	k := []string{"1", "2"}
+	a := []string{"3", "4"}
+
+	c := &client{
+		BFn: func() Builder {
+			return cmds.NewBuilder(cmds.NoSlot)
+		},
+		DoFn: func(ctx context.Context, cmd Completed) (resp RedisResult) {
+			return newResult(strmsg('+', "OK"), nil)
+		},
+		DoMultiFn: func(ctx context.Context, multi ...Completed) (resp []RedisResult) {
+			for _, cmd := range multi {
+				if reflect.DeepEqual(cmd.Commands(), []string{"EVAL", body, "2", "1", "2", "3", "4"}) {
+					resp = append(resp, newResult(strmsg('+', "OK"), nil))
+				}
+			}
+			return resp
+		},
+	}
+
+	script := NewLuaScriptNoSha(body)
+	if v, err := script.ExecMulti(context.Background(), c, LuaExec{Keys: k, Args: a})[0].ToString(); err != nil || v != "OK" {
+		t.Fatalf("ret mistmatch")
+	}
+}
+
 func TestNewLuaScriptExecMultiRo(t *testing.T) {
 	defer ShouldNotLeaked(SetupLeakDetection())
 	body := strconv.Itoa(rand.Int())
@@ -190,6 +282,36 @@ func TestNewLuaScriptExecMultiRo(t *testing.T) {
 	}
 
 	script := NewLuaScriptReadOnly(body)
+	if v, err := script.ExecMulti(context.Background(), c, LuaExec{Keys: k, Args: a})[0].ToString(); err != nil || v != "OK" {
+		t.Fatalf("ret mistmatch")
+	}
+}
+
+func TestNewLuaScriptExecMultiRoNoSha(t *testing.T) {
+	defer ShouldNotLeaked(SetupLeakDetection())
+	body := strconv.Itoa(rand.Int())
+
+	k := []string{"1", "2"}
+	a := []string{"3", "4"}
+
+	c := &client{
+		BFn: func() Builder {
+			return cmds.NewBuilder(cmds.NoSlot)
+		},
+		DoFn: func(ctx context.Context, cmd Completed) (resp RedisResult) {
+			return newResult(strmsg('+', "OK"), nil)
+		},
+		DoMultiFn: func(ctx context.Context, multi ...Completed) (resp []RedisResult) {
+			for _, cmd := range multi {
+				if reflect.DeepEqual(cmd.Commands(), []string{"EVAL_RO", body, "2", "1", "2", "3", "4"}) {
+					resp = append(resp, newResult(strmsg('+', "OK"), nil))
+				}
+			}
+			return resp
+		},
+	}
+
+	script := NewLuaScriptReadOnlyNoSha(body)
 	if v, err := script.ExecMulti(context.Background(), c, LuaExec{Keys: k, Args: a})[0].ToString(); err != nil || v != "OK" {
 		t.Fatalf("ret mistmatch")
 	}