Hello,

First of all, sorry for bringing some mess to rueidis 🙏  But generally this change can be useful for many other users too, especially those from enterprise environments.

[FIPS 140-3 Compliance](https://go.dev/doc/security/fips140) document, which is part of the Go 1.24 release defines runtime GODEBUG runtime toggles for [FIPS](https://en.wikipedia.org/wiki/Federal_Information_Processing_Standards) compliance.

SHA-1 is not a FIPS-approved algorithm, so using it in integrations with Redis complicates compliancy audit. And also Go runtime panics in case of using Redis Lua with `fips140=only` mode: https://cs.opensource.google/go/go/+/refs/tags/go1.24.2:src/crypto/sha1/sha1.go;l=273

This PR adds support to avoid SHA-1 usage by introducing a couple of new constructors for Lua scripts: `NewLuaScriptNoSha` and `NewLuaScriptReadOnlyNoSha`. This will allow `rueidis` users avoid SHA-1 usage when really required, without adding any FIPS knowledge to the library itself. Probably a better way to organize API exists.

Not using `EVALSHA` introduces performance overhead, I tried to run one of the benchmarks which uses Lua call:

```
name                           old time/op    new time/op    delta
RedisPresence_1Ch/rd_single-8    3.30µs ± 1%    4.88µs ± 1%  +47.84%  (p=0.000 n=9+10)

name                           old alloc/op   new alloc/op   delta
RedisPresence_1Ch/rd_single-8      638B ± 0%      640B ± 0%   +0.38%  (p=0.006 n=10+8)

name                           old allocs/op  new allocs/op  delta
RedisPresence_1Ch/rd_single-8      10.0 ± 0%      10.0 ± 0%     ~     (all equal)
```

So it's sth like 300k ops/sec -> 200k ops/sec for this specific script, but the compliance is priority here for enterprises to simplify FIPS audit.

I think ideally, this should be addressed by Redis – by introducing SHA-2 support for Lua scripting maybe. But given we now have many Redis-compatible storages the process will take a lot of time definitely. Linking to the related issue in Redis BTW - https://github.com/redis/redis/issues/12690